+++ Forge/src/main/java/fr/zadar/elementary/compat/jei/ElementsImprovementCategory.java
@@ -18,14 +18,14 @@ import org.jetbrains.annotations.Nullable;
 
 public class ElementsImprovementCategory implements IRecipeCategory<ElementsImprovementRecipe> {
     public static final ResourceLocation UID = new ResourceLocation(ElementaryForge.MOD_ID, "elements_improvement");
-    public static final ResourceLocation TEXTURE = new ResourceLocation(ElementaryForge.MOD_ID, "textures/gui/elements_improvement_gui.png");
+    public static final ResourceLocation TEXTURE = new ResourceLocation(ElementaryForge.MOD_ID, "textures/gui/compat/elements_improvement_gui_simplified.png");
     public static final RecipeType<ElementsImprovementRecipe> ELEMENTS_IMPROVEMENT_TYPE = new RecipeType<>(UID, ElementsImprovementRecipe.class);
 
     private final IDrawable background;
     private final IDrawable icon;
 
     public ElementsImprovementCategory(IGuiHelper helper) {
-        this.background = helper.createDrawable(TEXTURE, 50, 10, 90, 65);
+        this.background = helper.createDrawable(TEXTURE, 0, 0, 97, 67);
         this.icon = helper.createDrawableIngredient(VanillaTypes.ITEM_STACK, new ItemStack(ModBlocks.ELEMENTS_IMPROVEMENT.get()));
     }
 
@@ -51,8 +51,8 @@ public class ElementsImprovementCategory implements IRecipeCategory<ElementsImpr
 
     @Override
     public void setRecipe(IRecipeLayoutBuilder builder, ElementsImprovementRecipe recipe, IFocusGroup focuses) {
-        builder.addSlot(RecipeIngredientRole.INPUT, 6, 7).addIngredients(recipe.getIngredients().get(0));
-        builder.addSlot(RecipeIngredientRole.INPUT, 6, 43).addIngredients(recipe.getIngredients().get(1));
-        builder.addSlot(RecipeIngredientRole.OUTPUT, 66, 25).addItemStack(recipe.getResultItem(null));
+        builder.addSlot(RecipeIngredientRole.INPUT, 9, 8).addIngredients(recipe.getIngredients().get(0));
+        builder.addSlot(RecipeIngredientRole.INPUT, 9, 44).addIngredients(recipe.getIngredients().get(1));
+        builder.addSlot(RecipeIngredientRole.OUTPUT, 69, 26).addItemStack(recipe.getResultItem(null));
     }
 }