+++ Forge/src/main/java/fr/zadar/elementary/mixin/WeatherCommandMixin.java
@@ -1,0 +1,20 @@
+package fr.zadar.elementary.mixin;
+
+import fr.zadar.elementary.event.ToolEvents;
+import net.minecraft.commands.CommandSourceStack;
+import net.minecraft.server.commands.WeatherCommand;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+@Mixin(WeatherCommand.class)
+public class WeatherCommandMixin {
+    @Inject(method = "setClear", at = @At("HEAD"))
+    private static void onSetClear(CommandSourceStack sourceStack, int duration, CallbackInfoReturnable<Integer> cir) {
+        if (ToolEvents.isRaining) {
+            ToolEvents.isRaining = false;
+            ToolEvents.rainDuration = 0;
+        }
+    }
+}