+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [0.0.4] - [date]
+### Add
+Power for each element upgraded :
+- fire : the emberite pickaxe can smelt ores.
+- water : the aqualite hoe can be used on dirt to get a farmland block that doesn't need water and the sword can create rain.
+- wind : the aerilite tools can give jump boost 4 on right click
+### Fix
+- Crash from the aeroruptura model (fabric only)
+
 ## [0.0.3] - 21/06/2025
 ### Add
 - Air elements